UI Stuff

+++ Assets/Scripts/DamageOnImpact.cs
@@ -1,15 +1,13 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public class DamageOnImpact : MonoBehaviour
 {
-    [SerializeField] private float _damageOnImpact = 0f;
+    [SerializeField] private int _damageOnImpact = 0;
 
     private void OnCollisionEnter2D(Collision2D other)
     {
-        if(other.gameObject.TryGetComponent<Mortality>(out Mortality otherMortality)) {
+        if (other.gameObject.TryGetComponent<Mortality>(out Mortality otherMortality)) {
             otherMortality.damage(_damageOnImpact);
-        }       
+        }
     }
 }
